refactor(storage): extract proximadb-storage-transaction crate (root decomposition) (#329)
* refactor(storage): extract proximadb-storage-transaction crate

Move src/storage/transaction (multi-model ACID transaction coordinator:
manager, contexts, isolation levels, operations, write-ahead log) into a
dedicated workspace crate as a slice of the root-crate decomposition. The
root crate re-exports it as `crate::storage::transaction`, so all consumers
are untouched (zero call-site churn).

Pure git-mv + re-export; no behavioral change. Two test-module self-paths
rewritten crate-local (crate::operations::VectorOperation); serde_json/uuid
(v4)/tempfile deps declared for the leaf.

- root src/ shrinks by ~2.7K LOC across 6 files (one fewer root module)
- proximadb-storage-transaction: 35 tests pass, clippy -D warnings clean
- server/bench/migrate binaries build; layering green (storage tier, no
  upward edge); affected_crates maps the subtree to the leaf, not root

Ref TD-104, ROOT_CRATE_DECOMPOSITION_PLAN_2026_06_21.

* docs(design): update decomposition plan to verified develop reality

Record actual state of the storage-leaf slices on origin/develop:
- kv and transaction (this PR) are extracted; encryption is the next clean leaf
- formats and cache, the originally-named "quick wins", are both disproved as
  Low-risk: formats is engine-bound (adapters.rs -> all six engines); cache
  depends on the root-resident 90-consumer storage::traits::UnifiedStorageFormat
  plus root metrics/observability, so it is gated behind the D port-inversion

Adds a Status column to the Sequencing table and replaces "First action" with
a dated Status & next-action section.

Ref ROOT_CRATE_DECOMPOSITION_PLAN_2026_06_21, #145, #329.

---------
