+++ Cargo.toml
@@ -77,6 +77,7 @@ members = [
     "crates/storage/proximadb-storage-filesystem-types",
     "crates/storage/proximadb-storage-strategy",
     "crates/storage/proximadb-storage-tenant",
+    "crates/storage/proximadb-storage-transaction",
     "crates/query/proximadb-document-query",
     "crates/query/proximadb-graph-arrow",
     "crates/query/proximadb-graph-query",
@@ -215,6 +216,7 @@ proximadb-storage-optimization = { path = "crates/storage/proximadb-storage-opti
 proximadb-storage-filesystem-types = { path = "crates/storage/proximadb-storage-filesystem-types" }
 proximadb-storage-strategy = { path = "crates/storage/proximadb-storage-strategy" }
 proximadb-storage-tenant = { path = "crates/storage/proximadb-storage-tenant" }
+proximadb-storage-transaction = { path = "crates/storage/proximadb-storage-transaction" }
 crc32fast = "1.4"
 rmp-serde = "1.1"
 tonic-prost = "0.14"
@@ -306,6 +308,7 @@ proximadb-storage-optimization = { path = "crates/storage/proximadb-storage-opti
 proximadb-storage-filesystem-types = { path = "crates/storage/proximadb-storage-filesystem-types" }
 proximadb-storage-strategy = { path = "crates/storage/proximadb-storage-strategy" }
 proximadb-storage-tenant = { path = "crates/storage/proximadb-storage-tenant" }
+proximadb-storage-transaction = { path = "crates/storage/proximadb-storage-transaction" }
 proximadb-telemetry = { path = "crates/horizontal/proximadb-telemetry" }
 proximadb-filter-expression = { path = "crates/foundation/proximadb-filter-expression" }
 proximadb-cache = { path = "crates/foundation/proximadb-cache" }
