+++ crates/storage/proximadb-storage-transaction/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "proximadb-storage-transaction"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Multi-model ACID transaction coordinator (manager, contexts, isolation levels, operations, log) for ProximaDB"
+
+[lib]
+name = "proximadb_storage_transaction"
+path = "src/lib.rs"
+
+[dependencies]
+serde.workspace = true
+serde_json.workspace = true
+parking_lot.workspace = true
+tokio.workspace = true
+uuid = { version = "1", features = ["v4"] }
+proximadb-kernel.workspace = true
+
+[dev-dependencies]
+tempfile.workspace = true
+
+[lints]
+workspace = true
