+++ crates/storage/proximadb-storage-transaction/src/lib.rs
@@ -53,6 +53,10 @@
 //! // Commit atomically across all stores
 //! tx.commit().await?;
 //! ```
+//!
+//! Extracted from the root crate's `src/storage/transaction` as a slice of the root-crate
+//! decomposition (see `docs/12-design/ROOT_CRATE_DECOMPOSITION_PLAN_2026_06_21.adoc`).
+//! The root crate re-exports it as `crate::storage::transaction` for source compatibility.
 
 pub mod context;
 pub mod isolation;
