+++ crates/storage/proximadb-storage-transaction/src/operations.rs
