+++ docs/12-design/ROOT_CRATE_DECOMPOSITION_PLAN_2026_06_21.adoc
@@ -72,8 +72,10 @@ Existing seams that help:
 
 * `src/storage/trait_components/` — `StorageReader/Writer/Compactor/Scan/...` ports (engines
   are already pluggable).
-* `src/storage/formats/` — `StorageFormat` trait + `FormatRegistry` (mature, leaf-ish seam).
-* Extracted: `proximadb-{records,codec,block-format,storage-common,object-store,iceberg-engine}`.
+* `src/storage/formats/` — `StorageFormat` trait + `FormatRegistry` (looks leaf-ish, but
+  `adapters.rs` binds all six engines — *engine-bound*, deferred behind E; see Sequencing).
+* Extracted: `proximadb-{records,codec,block-format,storage-common,object-store,iceberg-engine,
+  storage-kv,storage-transaction}`.
 
 == Strategy
 
@@ -126,20 +128,27 @@ the compute kernels by the distance/quant/bridge seam — not per file.
 
 == Sequencing (incremental)
 
-[cols="1,3,1,1",options="header"]
+[cols="1,3,1,1,2",options="header"]
 |===
-| Slice | Scope | ~LOC | Risk
-| A | `proximadb-storage-formats` ← `src/storage/formats/` (re-export) | ~5K | Low
-| B | `proximadb-storage-cache` ← `src/storage/cache/` (wraps `proximadb-cache`) | ~11K | Low
-| C | Persistence: WAL + disk-manager port crates | ~? | Med
-| D | *Enabler:* `IndexManager` (+ distance, metadata) **port crate(s)**; refactor engine/compaction to inject ports | n/a | Med-High
-| E | Engines: `proximadb-{sst,nova,viper}-engine` (depend on ports from D) | large | High
-| F | Multimodel / document / entity / transaction stores | large | High
+| Slice | Scope | ~LOC | Risk | Status (2026-06-25)
+| A0 | `proximadb-storage-kv` ← `src/storage/kv/` (re-export) | ~0.1K | Low | ✅ Done
+| A1 | `proximadb-storage-transaction` ← `src/storage/transaction/` (re-export) | ~2.7K | Low | ✅ Done (PR #329)
+| A2 | `proximadb-storage-encryption` ← `src/storage/encryption/` | ~1.7K | Low | **Next clean leaf** — only 2 refs to `crate::storage::persistence` to resolve
+| — | `proximadb-storage-formats` ← `src/storage/formats/` | ~5K | ~~Low~~ → High | 🚫 *Blocked, not a quick win.* `formats/adapters.rs` binds all six engines (`crate::storage::engines::*`, 12 refs) — engine-bound. Defer behind E. (Disproves the original "formats-first"; see #145.)
+| — | `proximadb-storage-cache` ← `src/storage/cache/` (wraps `proximadb-cache`) | ~11K | ~~Low~~ → High | 🚫 *Blocked, not a quick win.* 11/27 files depend on the root-resident, 90-consumer `storage::traits::UnifiedStorageFormat` (+ root `metrics`/`observability`). Requires D (trait port-inversion) first.
+| C | Persistence: WAL + disk-manager port crates | ~? | Med | Pending
+| D | *Enabler:* `IndexManager` (+ distance, metadata) **port crate(s)**; refactor engine/compaction to inject ports | n/a | Med-High | Linchpin (also unblocks cache's `storage::traits` dep)
+| E | Engines: `proximadb-{sst,nova,viper}-engine` (depend on ports from D) | large | High | Coordinated window
+| F | Multimodel / document / entity / transaction stores | large | High | Coordinated window
 |===
 
-Slices A and B are independent quick wins that prove the re-export pattern and start
-shrinking the binary. D is the linchpin: the ~403 up-edges mean engines cannot move until
-the index/compute/metadata boundaries are ports. E/F are large and must be coordinated.
+The clean storage leaves (kv, transaction, encryption) are the independent quick wins that
+prove the re-export pattern and start shrinking the binary. *Note:* the originally-named
+"quick wins" (formats, cache) were disproved on closer inspection — both have hard upward
+edges into root-resident subsystems and are blocked behind D/E, not Low-risk. D is the
+linchpin: the ~403 up-edges mean engines cannot move until the index/compute/metadata
+boundaries are ports — and the same `storage::traits::UnifiedStorageFormat` port-inversion
+unblocks cache. E/F are large and must be coordinated.
 
 == Multi-session coordination protocol (mandatory)
 
@@ -172,8 +181,19 @@ and several edit `src/storage`. Large file-moves conflict violently with in-flig
 * coverage producible without exclusions (eventually yes),
 * CI compile wall-clock for the root crate (down).
 
-== First action
+== Status & next action (2026-06-25)
 
-Execute *Slice A* (`proximadb-storage-formats`) once a cold-subtree check confirms
-`src/storage/formats/` is not under active edit, as a pure `git mv` + re-export PR. It is the
-smallest, lowest-coupling proof of the pattern and starts the shrink.
+The proof-of-pattern slices have landed: `proximadb-storage-kv` (done) and
+`proximadb-storage-transaction` (PR #329) are both pure `git mv` + re-export extractions of
+verified clean leaves, layering-green with no call-site churn.
+
+A cold-subtree check disproved the originally-named first two slices: *formats* is
+engine-bound (`adapters.rs` → all six engines) and *cache* depends on the root-resident
+90-consumer `storage::traits::UnifiedStorageFormat` plus root `metrics`/`observability`.
+Neither is a mechanical quick win; both are deferred behind the *D* port-inversion / *E*
+engine extractions.
+
+*Next clean leaf:* `proximadb-storage-encryption` ← `src/storage/encryption/` (~1.7K LOC) —
+resolve its 2 references to `crate::storage::persistence`, then a pure `git mv` + re-export PR.
+After the leaf inventory is exhausted, *D* (the trait/index/compute port-inversion enabler)
+is the linchpin that unblocks both cache and the large engine slices.
